+++ CHANGES.rst
@@ -20,6 +20,10 @@ Changes
 - Migrate doctests to Python 3.8 [#261]
 - Migrate Python tests to pytest [#265]
 
+Fixes
+-----
+- Add additional check to tell if it's safe to redirect stdout/err [#270]
+
 
 2.4.0 (2021-09-30)
 ==================